third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-change-document-domain.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Validate that a frame can't change it's document.domain and enter the suborigin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<script>
document.domain = '127.0.0.1';
window.secret = 'Unchanged';
window.onmessage = function(event) {
  assert_equals(secret, 'Unchanged',
    'secret should not be touchable by the child in a suborigin.');
  assert_equals(event.data,
    'SecurityError: Blocked a frame with origin ' +
    '\"http-so://foobar.127.0.0.1:8000\" from accessing a cross-origin ' +
    'frame.');
  done();
};
</script>
<iframe src="resources/child-changes-document-domain.php?suborigin=foobar"></iframe>
</html>
